ensure we sync spaces when found later
This fixes a timing bug, where we might receive the information about an acter space
without knowing it is an acter space, causing it to be included in the to-sync-list
but then skipped at actual syncing. When we then see it later, and now know it is
an acter space, we didn't try it again.

This fixes that by removing any non-acter spaces from the known list.