Fix several state related News Editor bugs:
There was a misunderstanding in how the state is managed (stateful vs. stateless) that
lead to a premature optimization on behalf of the stateful texteditor by trying to
have its state recalculated the least often. This is now less selective and the
state recalculation checks are expanded instead. Fixing several bugs like color
changing problems.