Simplify Manager-driven patterns, remove belongs-to-processing
Before this change these models that have a specific Manager were still informing
the blongs-to-model itself leading to that a) being processed and b) the key being
issued as changed. Neither is necessary for manager-driven objects where we have an
explicit key to listen to. Thus this PR simplifies that code and removes the
ActerModel from them to prevent that unnecessary processing and issuing