Do not set the language as a local config unless the user explicitly did so
In the previous code, upon first start up we'd implicitly write the language according to the system
settings to a local configuration. If the user now changed their system language, we'd still
force the app to start in the previous system language, despite the user never having set that
language as their preference. With this change we will not implictly _write_ a local configuration
but adhere to the system language until the user explicit set a language themselves