Drop containsKey for secure sessions restoring, more verbose logs
Facing #1541 and #1422 we suspect secure storage on latest iOS and
Sonoma is not working correctly under all circumstances. This commit
adds a bunch of rather verbose logging in that section to allow us to
idenitfy better where it comes from. Additionally, it drops the somewhat
quirky "containsKey" in favor of just trying to "read" immediately.
Maybe that already does it for us ...